+++ Libplanet/Action/IActionEvaluation.cs
@@ -26,7 +26,7 @@ namespace Libplanet.Action
         /// <summary>
         /// The result states that <see cref="Action"/> makes.
         /// </summary>
-        public IAccountStateDelta OutputStates { get; }
+        public IAccountStateDelta OutputState { get; }
 
         /// <summary>
         /// An exception that had risen during evaluation.